add locality-checker
Locality checker is a container for detecting the region and availability zone
of a CockroachDB pod running in a public Kubernetes cloud offering.

It is meant to be run as an [init container](https://kubernetes.io/docs/concepts/workloads/pods/init-containers/),
which writes the region and zone values of the pod to a volume mounted at
`/etc/cockroach-locality`. An example Kubernetes manifest which uses the
container is provided.